requirements.txt
fastapi
uvicorn
Pillow
requests
beautifulsoup4
pydantic
python-dotenv
moviepy
dropbox
